Transform the methods into helper functions
