+++ astropy/visualization/wcsaxes/core.py
@@ -8,7 +8,6 @@ from matplotlib import rcParams
 from matplotlib.artist import Artist
 from matplotlib.axes import Axes, subplot_class_factory
 from matplotlib.transforms import Affine2D, Bbox, Transform
-from mpl_toolkits.axes_grid1.anchored_artists import AnchoredEllipse, AnchoredSizeBar
 
 import astropy.units as u
 from astropy.coordinates import BaseCoordinateFrame, SkyCoord
@@ -16,7 +15,6 @@ from astropy.utils import minversion
 from astropy.utils.compat.optional_deps import HAS_PIL
 from astropy.wcs import WCS
 from astropy.wcs.wcsapi import BaseHighLevelWCS, BaseLowLevelWCS
-from astropy.wcs.utils import proj_plane_pixel_scales
 
 from .coordinates_map import CoordinatesMap
 from .frame import RectangularFrame, RectangularFrame1D
@@ -27,16 +25,6 @@ from .wcsapi import IDENTITY, transform_coord_meta_from_wcs
 __all__ = ['WCSAxes', 'WCSAxesSubplot']
 
 VISUAL_PROPERTIES = ['facecolor', 'edgecolor', 'linewidth', 'alpha', 'linestyle']
-CORNERS = {
-    "top right": 1,
-    "top left": 2,
-    "bottom left": 3,
-    "bottom right": 4,
-    "right": 5,
-    "left": 6,
-    "bottom": 8,
-    "top": 9,
-}
 
 
 class _WCSAxesArtist(Artist):
@@ -371,156 +359,6 @@ class WCSAxes(Axes):
         args, kwargs = self._transform_plot_args(*args, **kwargs)
 
         return super().scatter(*args, **kwargs)
-
-    def add_beam(
-        self,
-        header=None,
-        major=None,
-        minor=None,
-        angle=None,
-        corner="bottom left",
-        frame=False,
-        borderpad=0.4,
-        pad=0.5,
-        **kwargs,
-    ):
-        """
-        Display the beam shape and size
-
-        By default, this method will search for the ``BMAJ``, ``BMIN``, and ``BPA``
-        keywords in the FITS header to set the major and minor axes and the
-        position angle on the sky.
-
-        Parameters
-        ----------
-        header : :class:`~astropy.io.fits.Header`, optional
-            Header containing the beam parameters.
-        major : float or :class:`~astropy.units.Quantity`, optional
-            Major axis of the beam in degrees or an angular quantity (overrides
-            ``BMAJ`` if present in header).
-        minor : float, or :class:`~astropy.units.Quantity`, optional
-            Minor axis of the beam in degrees or an angular quantity (overrides
-            ``BMIN`` if present in header).
-        angle : float or :class:`~astropy.units.Quantity`, optional
-            Position angle of the beam on the sky in degrees or an angular
-            quantity (overrides ``BPA`` if present in header) in the anticlockwise
-            direction.
-        corner : str, optional
-            The beam location. Acceptable values are ``'left'``, ``'right'``,
-            ``'top'``, 'bottom', ``'top left'``, ``'top right'``, ``'bottom left'``
-            (default), and ``'bottom right'``.
-        frame : bool, optional
-            Whether to display a frame behind the beam (default is ``False``).
-        borderpad : float, optional
-            Border padding, in fraction of the font size. Default is 0.4.
-        pad : float, optional
-            Padding around the beam, in fraction of the font size. Default is 0.5.
-        kwargs
-            Additional arguments are passed to :class:`matplotlib.patches.Ellipse`.
-        """
-
-        if header:
-            major = header["BMAJ"]
-            minor = header["BMIN"]
-            angle = header["BPA"]
-
-        if isinstance(major, u.Quantity):
-            major = major.to(u.degree).value
-
-        if isinstance(minor, u.Quantity):
-            minor = minor.to(u.degree).value
-
-        if isinstance(angle, u.Quantity):
-            angle = angle.to(u.degree).value
-
-        if self.wcs.is_celestial:
-            pix_scale = proj_plane_pixel_scales(self.wcs)
-            sx = pix_scale[0]
-            sy = pix_scale[1]
-            degrees_per_pixel = np.sqrt(sx * sy)
-        else:
-            raise ValueError("Cannot show beam when WCS is not celestial")
-
-        minor /= degrees_per_pixel
-        major /= degrees_per_pixel
-
-        corner = CORNERS[corner]
-
-        beam = AnchoredEllipse(
-            self.transData,
-            width=minor,
-            height=major,
-            angle=angle,
-            loc=corner,
-            pad=pad,
-            borderpad=borderpad,
-            frameon=frame,
-        )
-        beam.ellipse.set(**kwargs)
-
-        self.add_artist(beam)
-
-    def add_scalebar(
-        self,
-        length,
-        label=None,
-        corner="bottom right",
-        frame=False,
-        borderpad=0.4,
-        pad=0.5,
-        **kwargs,
-    ):
-        """Add a scale bar
-
-        Parameters
-        ----------
-        length : float or :class:`~astropy.units.Quantity`
-            The lenght of the scalebar in degrees or an angular quantity
-        label : str, optional
-            Label to place below the scale bar
-        corner : str, optional
-            Where to place the scale bar. Acceptable values are:, ``'left'``,
-            ``'right'``, ``'top'``, ``'bottom'``, ``'top left'``, ``'top right'``,
-            ``'bottom left'`` and ``'bottom right'`` (default)
-        frame : bool, optional
-            Whether to display a frame behind the scale bar (default is ``False``)
-        borderpad : float, optional
-            Border padding, in fraction of the font size. Default is 0.4.
-        pad : float, optional
-            Padding around the scale bar, in fraction of the font size. Default is 0.5.
-        kwargs
-            Additional arguments are passed to :class:`mpl_toolkits.axes_grid1.anchored_artists.AnchoredSizeBar`.
-
-        """
-
-        if isinstance(length, u.Quantity):
-            length = length.to(u.degree).value
-
-        if self.wcs.is_celestial:
-            pix_scale = proj_plane_pixel_scales(self.wcs)
-            sx = pix_scale[0]
-            sy = pix_scale[1]
-            degrees_per_pixel = np.sqrt(sx * sy)
-        else:
-            raise ValueError("Cannot show scalebar when WCS is not celestial")
-
-        length = length / degrees_per_pixel
-
-        corner = CORNERS[corner]
-
-        scalebar = AnchoredSizeBar(
-            self.transData,
-            length,
-            label,
-            corner,
-            pad=pad,
-            borderpad=borderpad,
-            sep=5,
-            frameon=frame,
-            **kwargs,
-        )
-
-        self.add_artist(scalebar)
 
     def reset_wcs(self, wcs=None, slices=None, transform=None, coord_meta=None):
         """
